Remove the rubocop rule 'ClassAndModuleChildren'
