+++ .rubocop.yml
@@ -26,10 +26,6 @@ Metrics/AbcSize:
   Enabled: true
   Max: 20
 
-# ModuleとClassの形式をcompactに設定
-Style/ClassAndModuleChildren:
-  EnforcedStyle: compact
-
 # モジュールの行数制限についてテストファイルを除外
 Metrics/BlockLength:
   Exclude:
