+++ app/controllers/admin/users_controller.rb
@@ -1,17 +1,19 @@
 # frozen_string_literal: true
 
-class Admin::UsersController < Admin::BaseController
-  def index
-    fetcher = User::Fetcher.new
-    fetcher.find_all_by(users_params)
-    render json: { list: fetcher.users, max_page: fetcher.max_page },
-           include: :admin,
-           methods: %i[active current_sign_in_at], status: :ok
-  end
+module Admin
+  class UsersController < Admin::BaseController
+    def index
+      fetcher = User::Fetcher.new
+      fetcher.find_all_by(users_params)
+      render json: { list: fetcher.users, max_page: fetcher.max_page },
+             include: :admin,
+             methods: %i[active current_sign_in_at], status: :ok
+    end
 
-  private
+    private
 
-  def users_params
-    params.permit(:page)
+    def users_params
+      params.permit(:page)
+    end
   end
 end
